fix git provider checkout of a remote ref on an existing repo
Per discussion of https://github.com/puppetlabs/puppetlabs-vcsrepo/issues/51 in
the git channel on freenode, EugeneKay <eugene@kashpureff.org> stated that `git
rev-parse` is not capable of inspecting remote refs but that `git ls-remote`
is.  This patch makes a second attempt to resolve the ref with `ls-remote` if
`rev-parse` fails.

The git provider also appears to support several type features that are not
tagged under `has_features`.  It's not clear if this is the best way to resolve
this issue or if the provider should be refactored to work with different type
features.

Demonstration of the problem with changing refs (branches and tags)

    $ git --version
    git version 1.7.1
    $ cat master.pp branch.pp
      vcsrepo { '/tmp/vcsrepo':
        ensure    => present,
        provider  => git,
        source    => 'https://github.com/puppetlabs/puppetlabs-vcsrepo.git',
        revision  => 'master',
      }
      vcsrepo { '/tmp/vcsrepo':
        ensure    => present,
        provider  => git,
        source    => 'https://github.com/puppetlabs/puppetlabs-vcsrepo.git',
        revision  => 'feature/cvs',
      }
    $ puppet apply --modulepath=`pwd`/.. master.pp
    Notice: /Stage[main]//Vcsrepo[/tmp/vcsrepo]/ensure: Creating repository from present
    Notice: /Stage[main]//Vcsrepo[/tmp/vcsrepo]/ensure: created
    Notice: Finished catalog run in 2.19 seconds
    $ puppet apply --modulepath=`pwd`/.. branch.pp
    Error: /Stage[main]//Vcsrepo[/tmp/vcsrepo]: Could not evaluate: Execution of '/usr/bin/git rev-parse feature/cvs' returned 128: fatal: ambiguous argument 'feature/cvs': unknown revision or path not in the working tree.
    Use '--' to separate paths from revisions
    feature/cvs

    Notice: Finished catalog run in 1.69 seconds