Add acceptance tests for git protocols using clone
The git_clone_protocols_spec.rb will be used to enumerate over the
available protocols for git to use to clone to disk. Support
key files are also included for https/ssl use.

Protocols tested

    * file protocol (raw filepath)
    * file protocol (file://)
    * git protocol (git://)
    * http protocol (http://)
    * https protocol (https://)
    * ssh protocol (ssh://user@host)

TODO: Currently, the ssl cert is not added to the CA on the host.
This causes the git clone to fail because the cert is not trusted
and the client does not proceed with a non-trusted ssl connection.