(QENG-798) Update acceptance helpers to create distmoduledir
Previously we were creating the module dir on foss installs, but not on PE.
PE agents do not have a module dir deployed as part of the base PE install,
so to test on PE agents we need to create the module dir. `mkdir -p` is
idempotent, so we use it as a cudgle here. We also update the beaker_helper
and spec_helper_acceptance to contain the same installation steps for test
dependencies, warnings, and installation methods.