fix branch existence determintaion functionality
this will stop failing in case there is a tag which is a substring of an existing branch name.
for example if there is a tag 'release' and a branch 'release/integration' current code will match 'release/integration' with pattern 'release' and decide that a branch checkout is needed. but release branch does not exist so it will fail.
this simple fix resolves the issue.