"Better" handling of :ensure changes
Currently the git provider handles changes between :bare/:mirror and
:present/:latest inside its checkout and init calls. That's a bit silly,
so I moved the logic inside the ensurable block. Now when you change the
ensure type it will show the proper "ensure changed" notice.

This meant I had to pull the conversion functions up from the private
section of the methods to the public.

I also fixed up some of the code to be more robust. It now sets mirror
properties and all that.